Add ability to verify the signature of messages
If the browser has implemented the WebCrypto API and ECDSA with P-256
curve is available a textarea and a button will be shown in the notes
column next to the signature of the message. The signature of the
message will be checked using the supplied digital signature public key
in PEM format or the digital signature certificate in X.509 PEM format.